# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger",
        "description": "TeaEirro",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/getTea/{id}": {
            "get": {
                "tags": [
                    "Tea"
                ],
                "operationId": "2d35267bb798978448b78db18bfa1f8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tea by ID."
                    },
                    "404": {
                        "description": "Tea not found."
                    }
                }
            }
        },
...
        "/api/addTea": {
            "post": {
                "tags": [
                    "Tea"
                ],
                "operationId": "2ba83c379af21cb089a7bab640c01968",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description",
                                    "type_id",
                                    "origin_id",
                                    "ingredients",
                                    "images[]"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "origin_id": {
                                        "type": "integer"
                                    },
                                    "ingredients": {
                                        "type": "string"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Tea."
                    }
                }
            }
        },
        "/api/deleteTea/{id}": {
            "delete": {
                "tags": [
                    "Tea"
                ],
                "operationId": "3290ed36cb47749e94787d8c203096ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tea to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tea and images deleted successfully."
                    },
                    "404": {
                        "description": "Tea not found."
                    },
                    "500": {
                        "description": "Error deleting tea."
                    }
                }
            }
        },
        "/api/editTea/{id}": {
            "post": {
                "tags": [
                    "Tea"
                ],
                "operationId": "ba2d48de4dc36bbfbb1f5ff26eab3637",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tea to edit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description",
                                    "type_id",
                                    "origin_id",
                                    "ingredients",
                                    "images[]"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "origin_id": {
                                        "type": "integer"
                                    },
                                    "ingredients": {
                                        "type": "string"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tea edited successfully."
                    },
                    "404": {
                        "description": "Tea not found."
                    },
                    "400": {
                        "description": "Validation error."
                    },
                    "500": {
                        "description": "Error editing tea."
                    }
                }
            }
        },
        "/api/deleteTeaImage/{teaId}/{imageName}": {
            "delete": {
                "tags": [
                    "Tea"
                ],
                "operationId": "403d44d8b874097511b9d7c36dc44a19",
                "parameters": [
                    {
                        "name": "teaId",
                        "in": "path",
                        "description": "ID of the tea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "imageName",
                        "in": "path",
                        "description": "Name of the image to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tea image deleted successfully."
                    },
                    "404": {
                        "description": "Tea or image not found."
                    },
                    "500": {
                        "description": "Error deleting tea image."
                    }
                }
            }
        },
        "/api/addTeaImage/{teaId}": {
            "post": {
                "tags": [
                    "Tea"
                ],
                "operationId": "d0ebb65f588406a25d1b9772af8dab4f",
                "parameters": [
                    {
                        "name": "teaId",
                        "in": "path",
                        "description": "ID of the tea to add an image to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tea image added successfully."
                    },
                    "404": {
                        "description": "Tea not found."
                    },
                    "400": {
                        "description": "Validation error."
                    },
                    "500": {
                        "description": "Error adding tea image."
                    }
                }
            }
        }
    }
}